+++ CHANGES.md
@@ -3,6 +3,9 @@
 - Added support for float arrays in ocaml-sys
 - Renamed `Value::float` to `Value::f64` and `Value::float_val` to `Value::f64_val`
 - Added `Value::alloc_f64_array`, `Value::f64_field` and `Value::store_f64_field`
+- Added `ocaml::import!` macro for calling OCaml functions from Rust
+- Added `ocaml::sig` proc-macro for generating `external` and type signatures
+- Added ocaml-build crate for generating OCaml code from `ocaml::sig` macros
 
 ## 0.22.4